+++ rusk-wallet/CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fix transaction history error when the wallet has no stake [#3734]
+- Fix transaction history fail after shield/unshield conversions [#3600]
 
 ## [0.2.0] - 2025-05-07
 
@@ -90,6 +91,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#3650]: https://github.com/dusk-network/rusk/issues/3650
 [#3623]: https://github.com/dusk-network/rusk/issues/3623
 [#3602]: https://github.com/dusk-network/rusk/issues/3602
+[#3600]: https://github.com/dusk-network/rusk/issues/3600
 [#3598]: https://github.com/dusk-network/rusk/issues/3598
 [#3597]: https://github.com/dusk-network/rusk/issues/3597
 [#3593]: https://github.com/dusk-network/rusk/issues/3593